[BEAM-7428] Output the timestamp on elements in ReadAllViaFileBasedSource
